film_hobo/requirement.txt
argcomplete==1.12.3
asgiref==3.2.10
attrs==21.2.0
autobahn==21.2.1
Automat==20.2.0
boto3==1.17.87
botocore==1.20.87
braintree==4.10.0
certifi==2021.5.30
cffi==1.14.5
cfn-flip==1.2.3
channels==3.0.3
chardet==4.0.0
click==8.0.1
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
DateTimeRange==1.0.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1
django-allauth==0.44.0
django-appconf==1.0.4
django-autoslug==1.9.8
django-bootstrap-datepicker-plus==3.0.5
django-braces==1.14.0
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-environ==0.4.5
django-filter==2.4.0
django-import-export==2.5.0
django-js-asset==1.2.2
django-npm==1.0.0
django-paypal==1.1.1
django-phonenumber-field==5.0.0
django-pipeline==2.0.6
django-rest-auth==0.9.5
django-rest-authemail==2.0.6
django-s3-storage==0.13.4
django-select2==7.7.1
django-solo==1.1.3
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
durationpy==0.5
environ==1.0
et-xmlfile==1.1.0
flake8==3.9.2
future==0.18.2
hjson==3.0.2
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
jmespath==0.10.0
jsmin==2.2.2
kappa==0.6.0
MarkupPy==1.14
mbstrdecoder==1.0.1
mccabe==0.6.1
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.7
packaging==20.9
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
pep517==0.10.0
phonenumbers==8.12.21
Pillow==8.2.0
pip-tools==6.1.0
placebo==0.9.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
reportlab==3.5.68
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.4.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.61.0
troposphere==2.7.1
Twisted==21.2.0
txaio==21.2.1
typepy==1.1.5
urllib3==1.26.5
Werkzeug==0.16.1
whitenoise==5.2.0
wsgi-request-logger==0.4.6
xlrd==2.0.1
xlwt==1.3.0
zappa==0.52.0
zappa-django-utils==0.4.1
zope.interface==5.4.0
